Move gnuplot and matplot functions to separate files, Kay suggestions
